+++ tests/integration/cases/call_per_element_false/Clojure_call.clj
@@ -1,0 +1,2 @@
+(defn process [& args] nil)
+(process :data 1)