+++ tests/integration/cases/call_scalar_args/Clojure_call.clj
@@ -1,0 +1,4 @@
+(defn process [& args] nil)
+(process :value "hello")
+(process :value 42)
+(process :value true)