+++ tests/integration/cases/call_transform_no_wrapper/Clojure_call.clj
@@ -1,0 +1,4 @@
+(defn process [& args] 0)
+(process :value "hello")
+(process :value 42)
+(process :value true)